refactor portfolio workspace routes
